requirements.txt
pyrogram
tgcrypto
Flask
python-dotenv